Dance Dance Revolution
A program that recreates an emulated version of the popular arcade game. Within this program I used multiple classes, as well as used a function create the motion of the arrows rotation upward on screen as well as a KeyPressed function to generate points and make the arrows disappear on the screen.